+++ hypervisor/include/arch/x86/irq.h
@@ -77,9 +77,6 @@
  */
 #define MAX_MSI_ENTRY 0x800U
 
-#define DEFAULT_DEST_MODE	IOAPIC_RTE_DESTMODE_LOGICAL
-#define DEFAULT_DELIVERY_MODE	IOAPIC_RTE_DELMODE_LOPRI
-
 #define INVALID_INTERRUPT_PIN	0xffffffffU
 
 /*
